second

+++ controller/controller.js
@@ -2,7 +2,7 @@ import userModel from "../models/usermodel.js";
 import blogModel from "../models/blogmodel.js";
 import jwt from "jsonwebtoken";
 import bcrypt from "bcryptjs";
-
+import mongoose from "mongoose";
 export const register=async(req,res)=>{
     try{
       const{name,emailid,age,phoneNumber,password}=req.body;
@@ -49,7 +49,7 @@ export const login=async(req,res)=>{
      }
      //generate jwt
 
-     const token=jwt.sign({_id:User._id},process.env.JWT_SECRET_KEY,{expiresIn:'1d'});
+     const token=jwt.sign({_id:User._id,role:User.role},process.env.JWT_SECRET_KEY,{expiresIn:'1d'});
      return res.status(200).json({User,token});
     }
     catch(error){
@@ -63,22 +63,23 @@ try{
   if(!title||!body){
     return res.status(400).json({message:"all feilds are required(title and body)"});
   }
-  const owner=req.user._id;
-  const existingBlog=await blogModel.findOne({$and:[{owner},{title}]});
+  const author=req.user._id;
+  const existingBlog=await blogModel.findOne({$and:[{author},{title}]});
   if(existingBlog){
     return res.status(400).json({message:"blog already exists for the user"});
   }
   const Blog=new blogModel({
     title,
     body,
-    author:owner,
+    author:author,
     tags
   });
   await Blog.save();
   return res.status(201).json({Blog});
 }
 catch(error){
-    return res.status(500).json({status:'fail',message:error.message})
+  if (!res.headersSent){return res.status(500).json({status:'fail',message:error.message})}
+    
 }
 };
 
@@ -102,37 +103,85 @@ export const likeBlog=async(req,res)=>{
     }
 };
 
-export const commentBlog=async(req,res)=>{
-    try{
-     const {_id}=req.params;
-     const{comment}=req.body;
-     if(!_id){
-        return res.status(400).json({message:"enter blog id in query url"});
-     }
-     const Blog=await blogModel.findById(_id);
-     if(!Blog){
-        return res.status(400).json({message:"Blog doesnt exist for the given Id"})
-     }
-     const user=req.user._id;
-     Blog.comments.push({ user: req.user._id, comment });
-     await Blog.save();
-     return res.status(200).json({Blog});
+export const commentBlog = async (req, res) => {
+  try {
+      const { _id } = req.params;   // Blog ID from URL params
+      const { comment } = req.body;  // Comment text from request body
+
+      if (!_id) {
+          return res.status(400).json({ message: "Enter blog id in query URL" });
+      }
+
+      const blog = await blogModel.findById(_id);
+      if (!blog) {
+          return res.status(400).json({ message: "Blog doesn't exist for the given ID" });
+      }
+
+      const user = req.user._id;  // Get the user ID from the authenticated user
+
+      // Check if the same user is commenting again and handle as per your requirements
+      blog.comments.push({ user, comment });
+      await blog.save();  // Save the blog with the new comment added
+
+      return res.status(200).json({ blog });
+  } catch (error) {
+      return res.status(500).json({ status: 'fail', message: error.message });
+  }
+};
+
+export const getCommentsOfaPost=async(req,res)=>{
+  try{
+    const {_id}=req.params;
+    if(!_id){
+      return res.status(400).json({message:"enter valid blog id"});
     }
-    catch(error){
-        return res.status(500).json({status:'fail',message:error.message})
+    const Blog=await blogModel.findById(_id);
+    if(!Blog){
+      return res.status(400).json({message:"Blog doesn't exist for the given Id"});
     }
+    return res.status(200).json(Blog.comments);
+  }
+  catch(error){
+      return res.status(500).json({status:'fail',message:error.message})
+  }
+  
+}
+
+export const singleCommentbyItsId =async(req,res)=>{
+  try{
+   const {_id}=req.params; //comment id
+   const {id}=req.query;   //blog id
+   if(!_id){
+    return res.status(400).json("enter valid comment id");
+   }
+   const blog = await blogModel.findById(id);
+if (!blog) {
+    return res.status(400).json({ message: "Blog doesn't exist for the given Id" });
+}
+const comment = blog.comments.find(c => c._id.toString() === _id);
+if (!comment) {
+    return res.status(404).json({ message: "Comment not found" });
+}
+return res.status(200).json(comment);
+  }
+  catch(error){
+    return res.status(500).json({message:error.message})
+  }
 }
 
 export const deleteBlog=async(req,res)=>{
     try{
+    const{_id:userId}=req.user;
     const{_id}=req.params;
     if(!_id){
         return res.status(400).json({message:"enter blog id in the url query"});
     }
-    const Blog=await blogModel.findByIdAndDelete(_id);
-    if(!Blog){
-        return res.status(400).json({message:"Blog doesnt exist for the given Id"})
+    const blog=await blogModel.findOne({_id,author:userId});
+    if(!blog){
+      return res.status(400).json({message:"blog id is not valid or ur not a blog author"});
     }
+    await blogModel.findByIdAndDelete(_id);
+  
    return res.status(200).json({message:"Blog removed"})
     }
     catch(error){
@@ -142,10 +191,120 @@ export const deleteBlog=async(req,res)=>{
 
 export const getBlogs=async(req,res)=>{
     try{
-     const blogs=await blogModel.find({});
+     const blogs=await blogModel.find({}).populate('author','name');
      return res.status(200).json({blogs});
     }
     catch(error){
         return res.status(500).json({status:'fail',message:error.message})
     }
+}
+
+export const blogofParticularUser=async(req,res)=>{
+  try{
+  const {_id}=req.params;
+  const Blog=await blogModel.find({author:_id}).populate('author','name');
+  return res.status(200).json({Blog});
+  }
+  catch(error){
+    return res.status(500).json({status:'fail',message:error.message}) 
+}
+}
+
+export const deleteAllBlogsByAdmin=async(req,res)=>{
+  try{
+    await blogModel.deleteMany({});
+    return res.status(200).json({message:"deleted all"});
+  }
+  catch(error){
+    return res.status(500).json({status:'fail',message:error.message}) 
+}
+}
+
+export const deletecomment=async(req,res)=>{
+  try{
+  const {_id}=req.params;   //comment id
+  const blogid=req.query.id;   //blog id
+  if (!mongoose.Types.ObjectId.isValid(_id) || !mongoose.Types.ObjectId.isValid(blogid)) {
+    return res.status(400).json({ message: "Comment ID and Blog ID are required." });
+}
+ const blog=await blogModel.findById(blogid);
+ if(!blog){
+  return res.status(400).json({message:'blog not exists'});
+ }
+ const comment=blog.comments.id(_id);
+ if(!comment){
+  return res.status(400).json({message:"comment not exists"});
+ }
+ if(comment.user.toString() !== req.user._id.toString()){
+  return res.status(400).json({ message: "You are not authorized to delete this comment."})
+ }
+ blog.comments.pull(_id);
+ await blog.save();
+ return res.status(200).json({ message: "Comment deleted successfully." });
+  }
+  catch(error){
+    return res.status(500).json({status:'fail',message:error.message}) 
+  }
+}
+
+
+export const deleteallcommentsofaBlog=async(req,res)=>{
+  try{
+   const {_id}=req.params;
+   const blog=await blogModel.findById(_id);
+   if(!blog){
+    return res.status(400).json({message:"blog not exists with the id"});
+   }
+  blog.comments=[];
+   await blog.save();
+   return res.status(200).json({ message: "Comments deleted successfully." });
+  
+  }
+  catch(error){
+    return res.status(500).json({status:'fail',message:error.message}) 
+  }
+}
+
+export const updatePost=async(req,res)=>{
+  try{
+   const {newtitle,newbody,newtags}=req.body;
+   const{_id}=req.params;
+   const {_id:userId}=req.user;
+   const blog=await blogModel.findOne({_id,author:userId});
+   if(!blog){
+    return res.status(400).json({message:"blogid is invalid or the user is not owner of the blog"});
+   }
+    blog.title=newtitle;
+    blog.body=newbody;
+    blog.tags=newtags;
+    await blog.save();
+    return res.status(200).json({message:"Blog updated with new records"});
+  }
+  catch(error){
+    return res.status(500).json({status:'fail',message:error.message}) 
+  }
+}
+
+
+export const updateComment=async(req,res)=>{
+  try{
+   const {_id}=req.params;
+   const blogid=req.query.id;
+   const {_id:userId}=req.user;
+   const {newcomment}=req.body;
+   if(!_id||!blogid){
+    return res.status(400).json({message:"all feilds are required"});
+   }
+   const blog=await blogModel.findOne({_id:blogid,author:userId});
+   if(!blog){
+    return res.status(400).json({message:"Blog doesn't exist or ur not the author of the blog"});
+   }
+   const com=blog.comments.id(_id);
+   com.comment=newcomment;
+   await blog.save();
+   return res.status(200).json({ message: "Comment updated successfully.", blog });
+  }
+  catch(error){
+    return res.status(500).json({status:'fail',message:error.message}) 
+  }
 }
